Async dispatch works initially.
